+++ client/src/components/select/index.tsx
@@ -49,7 +49,7 @@ const Select: React.FC<Props> = ({
       onMenuOpen={() => setMenuOpen(true)}
       onBlur={() => setMenuOpen(false)}
       placeholder={window.lang.global.select}
-      noOptionsMessage={window.lang.global.no_options}
+      noOptionsMessage={() => window.lang.global.no_options}
     ></ReactSelect>
   );
 };